+++ src/Block/Renderer/ListBlockRenderer.php
@@ -37,12 +37,19 @@ class ListBlockRenderer implements BlockRendererInterface
         $listData = $block->getListData();
 
         $tag = $listData->type === ListBlock::TYPE_UNORDERED ? 'ul' : 'ol';
-        $attr = ($listData->start === null || $listData->start === 1) ?
-            [] : ['start' => (string) $listData->start];
+
+        $attrs = [];
+        foreach ($block->getData('attributes', []) as $key => $value) {
+            $attrs[$key] = $htmlRenderer->escape($value, true);
+        }
+
+        if ($listData->start !== null && $listData->start !== 1) {
+            $attrs['start'] = (string) $listData->start;
+        }
 
         return new HtmlElement(
             $tag,
-            $attr,
+            $attrs,
             $htmlRenderer->getOption('inner_separator', "\n") . $htmlRenderer->renderBlocks(
                 $block->getChildren(),
                 $block->isTight()